Extract SelloDigitalCertificado to SelloDigitalCertificadoValidatorTrait
